requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
pytest
pyproj
kml2geojson
rasterio
geopandas
fiona